refactors acc to review comments
