+++ skill-tree/src/main/java/com/RDS/skilltree/Endorsement/EndorsementController.java
@@ -11,6 +11,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import java.io.IOException;
@@ -18,6 +19,7 @@ import java.util.UUID;
 
 @RestController
 @RequestMapping("/v1/endorsements")
+@Validated
 @Slf4j
 @RequiredArgsConstructor
 public class EndorsementController {
