+++ skill-tree/src/main/resources/application.properties
@@ -9,4 +9,4 @@ API_V1_PREFIX=/api/v1
 spring.datasource.version=8.1.0
 management.endpoints.web.exposure.include=health,info,metrics
 logging.level.root=INFO
-endorsements.dummmy-data.path=dummy-data/endorsements/endorsements.json
+endorsements.dummmy-data.path=dummy-data/endorsements.json
